Schema upgrade fixes

+++ upgrade/this_branch/convert_mac.sql
@@ -1,4 +1,4 @@
-ALTER TABLE interface ADD new_mac INTEGER;
+ALTER TABLE interface ADD new_mac NUMBER(19);
 
 UPDATE interface SET new_mac = TO_NUMBER(REPLACE(mac, ':'), 'xxxxxxxxxxxx') WHERE mac IS NOT NULL;
 
@@ -9,12 +9,12 @@ ALTER TABLE interface DROP COLUMN mac;
 ALTER TABLE interface RENAME COLUMN new_mac TO mac;
 ALTER TABLE interface ADD CONSTRAINT interface_mac_ck CHECK (mac >= 0 AND mac < 281474976710656);
 
-ALTER TABLE observed_mac ADD new_mac_address INTEGER;
+ALTER TABLE observed_mac ADD new_mac_address NUMBER(19);
 
 UPDATE observed_mac SET new_mac_address = TO_NUMBER(REPLACE(mac_address, ':'), 'xxxxxxxxxxxx');
 
 ALTER TABLE observed_mac DROP CONSTRAINT observed_mac_mac_address_nn;
-ALTER TABLE observed_mac MODIFY (new_mac_address INTEGER CONSTRAINT observed_mac_mac_address_nn NOT NULL);
+ALTER TABLE observed_mac MODIFY (new_mac_address NUMBER(19) CONSTRAINT observed_mac_mac_address_nn NOT NULL);
 ALTER TABLE observed_mac DROP PRIMARY KEY DROP INDEX;
 ALTER TABLE observed_mac ADD CONSTRAINT observed_mac_pk PRIMARY KEY (network_device_id, port, new_mac_address);
 ALTER TABLE observed_mac DROP COLUMN mac_address;
